Fix compiler warning/error in r10428
git-svn-id: http://skia.googlecode.com/svn/trunk@10429 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ gm/blurquickreject.cpp
@@ -64,7 +64,7 @@ protected:
         canvas->translate(kBoxSize, kBoxSize);
         canvas->drawRect(clipRect, hairlinePaint);
         canvas->clipRect(clipRect);
-        for (int i = 0; i < SK_ARRAY_COUNT(blurRects); ++i) {
+        for (size_t i = 0; i < SK_ARRAY_COUNT(blurRects); ++i) {
             blurPaint.setColor(colors[i]);
             canvas->drawRect(blurRects[i], blurPaint);
             canvas->drawRect(blurRects[i], hairlinePaint);
